updated tests and docs
