+++ packages/evals/src/metrics/nlp/completeness/index.test.ts
@@ -1,3 +1,5 @@
+import { describe, it, expect, beforeEach } from '@jest/globals';
+
 import { CompletenessMetric } from './index';
 
 describe('CompletenessMetric', () => {
