+++ packages/evals/src/metrics/nlp/keyword-coverage/index.test.ts
@@ -1,4 +1,4 @@
-import { it, expect } from '@jest/globals';
+import { describe, it, expect } from '@jest/globals';
 
 import { KeywordCoverageMetric } from './index';
