+++ packages/evals/src/metrics/nlp/tone/index.test.ts
@@ -1,4 +1,4 @@
-import { it, expect } from '@jest/globals';
+import { describe, it, expect } from '@jest/globals';
 
 import { ToneConsistencyMetric } from './index';
